had to update it yesterday

+++ src/structures/list/doubly/doubly_node/doubly_node.h
@@ -11,7 +11,7 @@ struct doubly_node {
 	struct doubly_node *previous; // uh
 };
 
-struct doubly_node doubly_node_construct(void *data, int size);
+struct doubly_node doubly_node_construct(void *data, size_t size);
 void doubly_node_destruct(struct doubly_node *doubly_node);
 
 #endif
